Fix: last durationMs Float conversion

+++ app/src/main/java/com/aiphoneguardian/app/ui/viewmodel/ScannerViewModel.kt
@@ -41,7 +41,7 @@ class ScannerViewModel @Inject constructor(
                     _scanProgress.value = when {
                         result.durationMs < 20 -> result.durationMs.toFloat() * 5f
                         result.durationMs < 50 -> result.durationMs.toFloat() * 2f
-                        else -> result.durationMs.coerceIn(0f, 100f)
+                        else -> result.durationMs.toFloat().coerceIn(0f, 100f)
                     }
 
                     if (result.threats.isNotEmpty()) {
